# docs/swagger.yaml
components:
  schemas:
    api.ErrorResponse:
      properties:
        error:
          example: validation failed
          type: string
      type: object
    api.createDatasetRequest:
      properties:
        name:
          type: string
        source:
          type: string
        ttl_s:
          type: integer
      type: object
...
info:
  description: 'Zeus is the execution plane for faults-lab. Manteion issues

    commands to it. This document is the integration contract between

    the two. Run lifecycle state machine and cross-cutting concerns

    (dataset handshake, error envelope) are documented in

    docs/api-contract.md.'
  title: Zeus Load Generation API
  version: '1.0'
openapi: 3.1.0
paths:
...
  /datasets:
    get:
      description: 'Returns an envelope { "datasets": [...] } with summary fields
        per dataset

        (id, name, source, size_bytes, ttl_s, created_at). Use GET /datasets/{id}

        for full pool stats.'
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: envelope with datasets array
      summary: List datasets
      tags:
      - datasets
    post:
      description: 'Persist a dataset shell. Source is one of "upload" (NDJSON push),
        "inline"

        (synthesized by run-create from dataset_inline), or "cache_box_dump" (future).

        Defaults: source="upload", ttl_s=86400 (24 h). Rows are pushed separately
        via

        POST /datasets/{id}/upload.'
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.createDatasetRequest'
                description: dataset shell
                summary: body
        description: dataset shell
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: 'envelope: id, name, source, created_at, ttl_s'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: invalid JSON
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: dataset name conflict
      summary: Register dataset
      tags:
      - datasets
  /datasets/{id}:
    delete:
      description: Refuses with 409 when an active run still references the dataset.
      parameters:
      - description: Dataset ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '204':
          description: dataset deleted
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: dataset not found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: dataset is in use by an active run
      summary: Delete dataset
      tags:
      - datasets
    get:
      description: 'Returns an envelope with id, name, source, pool_stats (per-pool
        row_count,

        size_bytes, fields), size_bytes, ttl_s, created_at. Use GET /datasets/{id}/sample

        to inspect rows.'
      parameters:
      - description: Dataset ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: dataset metadata envelope
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: dataset not found
      summary: Get dataset
      tags:
      - datasets
  /datasets/{id}/sample:
    get:
      description: 'Returns up to limit rows (default 10, capped at 100) from the
        named pool.

        The response envelope is { "pool": "<name>", "rows": [...] }.'
      parameters:
      - description: Dataset ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Pool name to sample
        in: query
        name: pool
        required: true
        schema:
          type: string
      - description: Max rows to return (default 10, max 100)
        in: query
        name: limit
        schema:
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: sample envelope
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: missing pool query parameter
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: dataset or pool not found
      summary: Sample dataset rows
      tags:
      - datasets
  /datasets/{id}/upload:
    post:
      description: 'Streams NDJSON pool batches into the dataset. Each line is

        {"pool":"<name>","rows":[{...},...]}. The endpoint is lenient: per-line decode

        errors are collected and returned in the response body but do not abort the

        ingest. The response envelope contains ingested (per-pool counts), total

        (rows accepted), and an optional errors array.'
      parameters:
      - description: Dataset ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/x-ndjson:
            schema:
              type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: ingest result envelope
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: ingest error (e.g. dataset not found at ingest time)
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: dataset not found
      summary: Upload dataset rows (NDJSON)
      tags:
      - datasets
...
servers:
- description: Local dev (HTTP)
  url: http://localhost:8080/api/v1
- description: Local dev (HTTPS)
  url: https://localhost:8080/api/v1